requirements.txt
gradio
tensorflow
Pillow